backend/requirements.txt
fastapi
uvicorn
python-multipart
cryptography
python-dotenv
